Added getHistory function
Added a function that enables the History module to return the history of the table as an array.

Provides easier usability to developers looking to give users the ability to make several changes to a table and save the values to a database in a single run. 

As of right now the functions just returns the history array from the history module, but the array could be modified in the function to allow for easier readability and less traversing.